Move import_hook code into the __init__.py file
